Convert CSPSim over to using SimCore. Also fixed a few warnings that cropped up from reordering some of the headers.
git-svn-id: https://csp.svn.sourceforge.net/svnroot/csp/trunk@1256 7152ba15-5bd8-0310-b828-e896570a432b

+++ CSP/CSPSim/Source/Sky.cpp
@@ -25,7 +25,6 @@
 
 
 #include "Sky.h"
-#include "Tools.h"
 #include "CSPSim.h"
 #include "Config.h"
 #include "Colorspace.h"
@@ -34,6 +33,7 @@
 
 #include <cmath>
 
+#include <SimCore/Util/StringTools.h>
 #include <SimData/Date.h>
 #include <SimData/Path.h>
 #include <SimData/Object.h>
@@ -1098,8 +1098,8 @@ public:
 		float a = 0.0;
 		float da = 2.0 * simdata::PI / m_Segments;
 		// XXX
-		//float radius = std::min(1000000.0, std::max(1.5*clip, sqrt(2.0 * 6370000.0 * altitude)));
-		float radius = std::min(99500.0f, std::max(1.5f*clip, sqrt(2.0f * 6370000.0f * altitude)));
+		//float radius = std::min(1000000.0, std::max(1.5*clip, sqrtf(2.0 * 6370000.0 * altitude)));
+		float radius = std::min(99500.0f, std::max(1.5f*clip, sqrtf(2.0f * 6370000.0f * altitude)));
 		// variation less than 2 pixels under most conditions:
 		if (fabs(radius - m_Radius) < 0.0005 * m_Radius) return;
 		//std::cerr << "HORIZON -> " << radius << "\n";
